refactor: reorganize PostRetrieval structure and create experiment design
- Move reinforce from reinforce/ to augment/reinforce.py
- Update registry to register augment.reinforce
- Create result_optimization_strategy/ directory
- Add PostRetrieval_Experiment_Design.md (569 lines)
- Update MemoryBank config with refactoring notes
