+++ packages/sage-benchmark/src/sage/benchmark/benchmark_memory/experiment/libs/post_retrieval/registry.py
@@ -4,6 +4,7 @@
 """
 
 from .augment import AugmentAction
+from .augment.reinforce import ReinforceAction
 from .base import BasePostRetrievalAction
 from .filter.threshold import ThresholdFilterAction
 from .filter.token_budget import TokenBudgetFilterAction
@@ -13,7 +14,6 @@ from .merge.multi_query import MultiQueryMergeAction
 from .merge.multi_tier import MultiTierMergeAction
 from .merge.scm_three_way import SCMThreeWayMergeAction
 from .none_action import NoneAction
-from .reinforce import ReinforceAction
 from .rerank.ppr import PPRRerankAction
 from .rerank.semantic import SemanticRerankAction
 from .rerank.time_weighted import TimeWeightedRerankAction
@@ -37,7 +37,8 @@ class PostRetrievalActionRegistry:
     - merge.link_expand: 链接扩展合并
     - merge.multi_query: 多查询合并
     - merge.multi_tier: 多层融合（MemGPT）
-    - augment: 结果增强
+    - augment: 结果增强（添加 persona/traits/summary）
+    - augment.reinforce: 记忆强化（更新记忆强度，MemoryBank）
     """
 
     _actions: dict[str, type[BasePostRetrievalAction]] = {}
@@ -114,8 +115,6 @@ PostRetrievalActionRegistry.register("merge.multi_query", MultiQueryMergeAction)
 PostRetrievalActionRegistry.register("merge.multi_tier", MultiTierMergeAction)
 PostRetrievalActionRegistry.register("scm_three_way", SCMThreeWayMergeAction)
 
-# Augment Action
+# Augment Actions
 PostRetrievalActionRegistry.register("augment", AugmentAction)
-
-# Reinforce Action (MemoryBank)
-PostRetrievalActionRegistry.register("reinforce", ReinforceAction)
+PostRetrievalActionRegistry.register("augment.reinforce", ReinforceAction)
