--- packages/sage-benchmark/src/sage/benchmark/benchmark_memory/experiment/libs/post_retrieval/reinforce/__init__.py
@@ -1,8 +1,0 @@
-"""Memory Reinforcement Actions
-
-实现 MemoryBank 论文的检索强化机制。
-"""
-
-from .base import ReinforceAction
-
-__all__ = ["ReinforceAction"]
